singularize ability names
