+++ app/abilities/article.js
@@ -5,6 +5,16 @@ import { Ability } from 'ember-can';
 
 export default Ability.extend({
   session: service(),
+  canCreate: computed('session.currentUser', function() {
+    return this.session.hasPermission('article.create');
+  }),
+  canShow: true,
+  canDestroy: computed('session.currentUser', function() {
+    return this.session.hasPermission('article.destroy');
+  }),
+  canSelectAllGroups: computed('session.currentUser', function() {
+    return this.session.hasPermission('article.update');
+  }),
   canShowArticleComments: computed('session.currentUser', 'model', function() {
     return this.session.hasPermission('article-comment.read') || this.get('model.publiclyVisible');
   }),
