+++ app/abilities/board-room-presence.js
