+++ app/abilities/form/response.js
@@ -5,6 +5,12 @@ import { Ability } from 'ember-can';
 
 export default Ability.extend({
   session: service(),
+  canShow: computed('session.currentUser', function() {
+    return this.session.hasPermission('form/response.read');
+  }),
+  canCreate: computed('session.currentUser', function() {
+    return this.session.hasPermission('form/response.create');
+  }),
   canDestroy: computed('session.currentUser', 'model', function() {
     return this.session.hasPermission('form/response.destroy') || this.isResponseOwner(this.model);
   }),
