+++ app/abilities/photo-album.js
