+++ app/abilities/poll.js
@@ -5,6 +5,15 @@ import { Ability } from 'ember-can';
 
 export default Ability.extend({
   session: service(),
+  canShow: computed('session.currentUser', function() {
+    return this.session.hasPermission('poll.read');
+  }),
+  canCreate: computed('session.currentUser', function() {
+    return this.session.hasPermission('poll.create');
+  }),
+  canDestroy: computed('session.currentUser', function() {
+    return this.session.hasPermission('poll.destroy');
+  }),
   canEdit: computed('session.currentUser', 'model', function() {
     return this.session.hasPermission('poll.update') || this.isPollOwner(this.model);
   }),
