+++ app/abilities/static-page.js
