+++ package.json
@@ -32,7 +32,7 @@
     "ember-ajax": "^3.0.0",
     "ember-auto-import": "^1.5.3",
     "ember-bootstrap": "2.6.0",
-    "ember-can": "^1.1",
+    "ember-can": "^2.0",
     "ember-cli": "~3.2.0",
     "ember-cli-babel": "^7.13",
     "ember-cli-content-security-policy": "^1.1.1",
